+++ applications/fishing-map/src/features/workspace/vessels/VesselLayerPanel.tsx
@@ -2,7 +2,6 @@ import React, { Fragment, useCallback, useState } from 'react'
 import cx from 'classnames'
 import { useSelector } from 'react-redux'
 import { useTranslation } from 'react-i18next'
-import { fitBounds } from 'viewport-mercator-project'
 import { DatasetTypes, Vessel } from '@globalfishingwatch/api-types'
 import { Switch, IconButton, Tooltip, ColorBar } from '@globalfishingwatch/ui-components'
 import {
@@ -15,7 +14,7 @@ import {
   filterSegmentsByTimerange,
 } from '@globalfishingwatch/data-transforms'
 import { formatInfoField } from 'utils/info'
-import { UrlDataviewInstance } from 'types'
+import { Bbox, UrlDataviewInstance } from 'types'
 import { AsyncReducerStatus } from 'utils/async-slice'
 import styles from 'features/workspace/shared/LayerPanel.module.css'
 import { useDataviewInstancesConnect } from 'features/workspace/workspace.hook'
@@ -23,8 +22,7 @@ import { resolveDataviewDatasetResource } from 'features/workspace/workspace.sel
 import { selectResourceByUrl } from 'features/resources/resources.slice'
 import I18nDate from 'features/i18n/i18nDate'
 import I18nFlag from 'features/i18n/i18nFlag'
-import { useMapboxInstance } from 'features/map/map.context'
-import useViewport from 'features/map/map-viewport.hooks'
+import { useMapFitBounds } from 'features/map/map-viewport.hooks'
 import { useTimerangeConnect } from 'features/timebar/timebar.hooks'
 import { DATAVIEW_INSTANCE_PREFIX } from 'features/dataviews/dataviews.utils'
 import ExpandedContainer from 'features/workspace/shared/ExpandedContainer'
@@ -51,8 +49,7 @@ type LayerPanelProps = {
 
 function LayerPanel({ dataview }: LayerPanelProps): React.ReactElement {
   const { t } = useTranslation()
-  const mapInstance = useMapboxInstance()
-  const { setMapCoordinates } = useViewport()
+  const fitBounds = useMapFitBounds()
   const { upsertDataviewInstance, deleteDataviewInstance } = useDataviewInstancesConnect()
   const { url } = resolveDataviewDatasetResource(dataview, { type: DatasetTypes.Vessels })
   const { url: trackUrl } = resolveDataviewDatasetResource(dataview, { type: DatasetTypes.Tracks })
@@ -66,25 +63,14 @@ function LayerPanel({ dataview }: LayerPanelProps): React.ReactElement {
     if (trackResource?.data) {
       const filteredSegments = filterSegmentsByTimerange(trackResource?.data, { start, end })
       const bbox = filteredSegments?.length ? segmentsToBbox(filteredSegments) : undefined
-      const { width, height } = (mapInstance as any)._canvas || {}
-      if (width && height && bbox) {
-        const [minLng, minLat, maxLng, maxLat] = bbox
-        const { latitude, longitude, zoom } = fitBounds({
-          bounds: [
-            [minLng, minLat],
-            [maxLng, maxLat],
-          ],
-          width,
-          height,
-          padding: 60,
-        })
-        setMapCoordinates({ latitude, longitude, zoom })
+      if (bbox) {
+        fitBounds(bbox as Bbox)
       } else {
         // TODO use prompt to ask user if wants to update the timerange to fit the track
         alert('The vessel has no activity in your selected timerange')
       }
     }
-  }, [mapInstance, setMapCoordinates, trackResource, start, end])
+  }, [fitBounds, trackResource, start, end])
 
   const layerActive = dataview?.config?.visible ?? true
   const onToggleLayerActive = () => {