+++ packages/react-hooks/src/use-tiles-loading/use-tiles-loading.ts
@@ -1,5 +1,6 @@
 import { useCallback, useState, useEffect } from 'react'
 import type { Map } from '@globalfishingwatch/mapbox-gl'
+
 function useTilesLoading(map?: Map) {
   const [loading, setLoading] = useState<boolean>(false)
 
@@ -11,18 +12,18 @@ function useTilesLoading(map?: Map) {
     setLoading(true)
   }, [])
 
-  const onLoadComplete = useCallback(() => {
-    setLoading(false)
-  }, [])
-
   useEffect(() => {
     if (map) {
       map.on('sourcedataloading', onLoad)
-      map.on('error', onLoadComplete)
-      map.on('sourcedata', onLoadComplete)
       map.on('idle', onIdle)
     }
-  }, [map, onIdle, onLoad, onLoadComplete])
+    return () => {
+      if (map) {
+        map.off('sourcedataloading', onLoad)
+        map.off('idle', onIdle)
+      }
+    }
+  }, [map, onIdle, onLoad])
 
   return loading
 }