+++ applications/amathea/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @globalfishingwatchapp/amathea
 
+## 0.14.6
+
+### Patch Changes
+
+- Updated dependencies [[`f7b017f0`](https://github.com/GlobalFishingWatch/frontend/commit/f7b017f0c72bd1f90e6a984ff5c26e524d7d735a)]:
+  - @globalfishingwatch/layer-composer@4.16.0
+  - @globalfishingwatch/react-hooks@4.9.3
+  - @globalfishingwatch/ui-components@2.14.1
+
 ## 0.14.5
 
 ### Patch Changes