+++ applications/fishing-map/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @globalfishingwatchapp/fishing-map
 
+## 0.27.7
+
+### Patch Changes
+
+- Updated dependencies [[`55796c7a`](https://github.com/GlobalFishingWatch/frontend/commit/55796c7a0a0b9972d3b7f8631566091dd0b74809), [`f7b017f0`](https://github.com/GlobalFishingWatch/frontend/commit/f7b017f0c72bd1f90e6a984ff5c26e524d7d735a)]:
+  - @globalfishingwatch/fourwings-aggregate@1.2.1
+  - @globalfishingwatch/layer-composer@4.16.0
+  - @globalfishingwatch/react-hooks@4.9.3
+  - @globalfishingwatch/ui-components@2.14.1
+
+## 0.27.6
+
+### Patch Changes
+
+- Updated dependencies [[`a607f131`](https://github.com/GlobalFishingWatch/frontend/commit/a607f13101520a9c1571a36cb8198b7af04d10ea)]:
+  - @globalfishingwatch/ocean-areas@0.1.2
+  - @globalfishingwatch/ui-components@2.14.0
+
+## 0.27.5
+
+### Patch Changes
+
+- fix useTilesLoading
+
+- Updated dependencies []:
+  - @globalfishingwatch/react-hooks@4.9.2
+
 ## 0.27.4
 
 ### Patch Changes